Refactor tasks import/export
Move ddln code from 'annotation' app to ddln package.
Move task import/export (persistence) code int their own packages.
Remove diverging implementation of ddln spotter dump.
Unify spotter and vls task workflow (especially regarding validation and reporting).
