+++ cvat/apps/annotation/ddln_vls.py
@@ -22,35 +22,36 @@ format_spec = {
 
 
 def dump(file_object, annotations):
-    import io
-    from cvat.apps.annotation.transports.csv import CsvZipExporter
-    from cvat.apps.annotation.transports.cvat import CVATImporter
-    from cvat.apps.annotation.transports.cvat.utils import FileLogger
-    from cvat.apps.engine.ddln.utils import write_task_mapping_file
+    from tempfile import TemporaryDirectory
+    from cvat.apps.dataset_manager.util import make_zip_archive
+    from cvat.apps.engine.ddln.transports import migrate, CsvDirectoryExporter, CsvDirectoryImporter, CVATImporter
+    from cvat.apps.engine.ddln.tasks.vls import VlsTaskHandler
+    from cvat.apps.engine.ddln.utils import write_task_mapping_file, DdlnYamlWriter
 
-    buffer = io.StringIO()
-    logger = FileLogger(buffer)
-    importer = CVATImporter(annotations, logger)
-    with CsvZipExporter(file_object) as exporter:
-        for frame_reader in importer.iterate_frames():
-            with exporter.begin_frame(frame_reader.name, frame_reader.sequence_name) as frame_writer:
-                for runway in frame_reader.iterate_runways():
-                    frame_writer.write_runway(runway)
-        zip_archive = exporter.get_archive()
-        zip_archive.writestr('validation.log', buffer.getvalue())
+    with TemporaryDirectory() as temp_dir:
+        importer = CVATImporter(annotations)
+        exporter = CsvDirectoryExporter(temp_dir)
+        handler = VlsTaskHandler()
+        migrate(importer, exporter, handler)
 
-        task_mapping_buffer = io.StringIO()
-        write_task_mapping_file(annotations._db_task, task_mapping_buffer)
-        zip_archive.writestr('task_mapping.csv', task_mapping_buffer.getvalue())
+        sequences = handler.load_sequences(CsvDirectoryImporter(temp_dir))
+        reporter = handler.validate(sequences)
+        validation_file = os.path.join(temp_dir, 'validation.txt')
+        reporter.write_text_report(open(validation_file, 'wt'), reporter.severity.WARNING)
+
+        yaml_writer = DdlnYamlWriter(annotations.meta['task']['name'], add_merger_info=False)
+        yaml_writer.write_metadata(open(os.path.join(temp_dir, "ddln.yaml"), 'w'))
+
+        task_mapping_filename = os.path.join(temp_dir, 'task_mapping.csv')
+        write_task_mapping_file(annotations._db_task, open(task_mapping_filename, 'wt'))
+        make_zip_archive(temp_dir, file_object)
 
 
 def load(file_object, annotations):
-    from cvat.apps.annotation.transports.cvat import CVATExporter
-    from cvat.apps.annotation.transports.csv import CsvZipImporter
+    from cvat.apps.engine.ddln.transports import migrate, CsvZipImporter, CVATExporter
+    from cvat.apps.engine.ddln.tasks.vls import VlsTaskHandler
 
     importer = CsvZipImporter(file_object)
-    with CVATExporter(annotations) as exporter:
-        for frame_reader in importer.iterate_frames():
-            with exporter.begin_frame(frame_reader.name, frame_reader.sequence_name) as frame_writer:
-                for runway in frame_reader.iterate_runways():
-                    frame_writer.write_runway(runway)
+    exporter = CVATExporter(annotations)
+    handler = VlsTaskHandler()
+    migrate(importer, exporter, handler)
