--- cvat/apps/annotation/transports/csv/__init__.py
@@ -1,2 +1,0 @@
-from .importer import CsvZipImporter, CsvDirectoryImporter
-from .exporter import CsvZipExporter, CsvDirectoryExporter
