--- cvat/apps/annotation/transports/cvat/__init__.py
@@ -1,2 +1,0 @@
-from .importer import CVATImporter
-from .exporter import CVATExporter
