+++ cvat/apps/engine/ddln/tasks/spotter/__init__.py
@@ -1,0 +1,1 @@
+from .handler import SpotterTaskHandler
