+++ cvat/apps/engine/ddln/tasks/utils.py
@@ -1,0 +1,2 @@
+def build_attrs_dict(shape):
+    return {a.name: a.value for a in shape.attributes}
